fix: bug in several versions of CopyFrom if the instance to copy from is identical to the destination add: physical unit system add: several new WPF controls, WPF support is now complete, so that AltaxoBase.WinForms was removed from the project
--HG--
extra : convert_revision : svn%3Ad1681fee-8e23-fc4c-b001-5eb8f207e95b/trunk%40693

+++ Altaxo/Base/Graph/Gdi/Plot/Styles/XYLineScatterPlotStyle.cs
@@ -249,7 +249,7 @@ namespace Altaxo.Graph.Gdi.Plot.Styles
     {
       get 
       {
-        return null==m_ScatterStyle ? 0 : m_ScatterStyle.SymbolSize;
+        return (float)( null==m_ScatterStyle ? 0 : m_ScatterStyle.SymbolSize);
       }
     }
